api/requirements.txt
tensorflow==2.10.1
tensorflow-serving-api==2.5.0
fastapi
uvicorn
python-multipart
pillow
numpy
matplotlib
